File uploads
Objective: To learn and analyze road accident data using Excel in order to identify patterns and trends that can be used to prevent future accidents.

Description: Cleaning of the data by removing any errors or inconsistencies.  Processing of the data by adding new columns, creating calculated fields, or grouping data by different categories .  Analysis of  the data using Excel's built-in tools, such as pivot tables and charts. Created charts and graphs to visualize the data and make it easy to understand and Created a dashboard that combines the data visualizations and makes it easy to interact with the data , Charts and graphs